requirements.txt
alpaca-py==0.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.5.7
boto3==1.26.36
botocore==1.29.36
certifi==2021.5.30
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==2.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==38.0.1
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.7
django-cors-headers==3.9.0
django-filter==21.1
django-guardian==2.4.0
django-storages==1.13.2
django-taggit==2.1.0
django-taggit-serializer==0.1.7
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.24.2
drf-spectacular-sidecar==2022.10.1
drf-yasg==1.21.4
Faker==15.1.0
finnhub-python==2.4.15
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.2
importlib-resources==5.10.0
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.16.0
MarkupSafe==2.1.1
msgpack==1.0.4
newsdataapi==0.1.5
numpy==1.21.5
oauthlib==3.2.2
packaging==21.3
pandas==1.3.5
Pillow==8.3.2
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
PyJWT==1.7.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.1
PyYAML==6.0
rave-python==1.2.16
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
sseclient-py==1.7.2
toml==0.10.2
Twisted==22.8.0
txaio==22.2.1
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.7
websockets==10.4
whitenoise==6.0.0
wrapt==1.14.1
zipp==3.9.0
zope.interface==5.5.0
